+++ CHANGES.rst
@@ -34,6 +34,7 @@ Improvements
   (:pr:`5936`, thanks :user:`openprojects`)
 - Show only active registration counts on the registration form management dashboard, and add
   an inactive registration count to the registration list (:pr:`5990`)
+- Store creation date of users and show it to admins (:pr:`5957`, thanks :user:`vasantvohra`)
 
 Bugfixes
 ^^^^^^^^